im dumb
main launch sequence was broken because the skipthrough was if dire was '.', this is dumb for many reasons, now it's been changed to 'GO!", which no directory will be named, because its literally missing a drive, and the only other thing it could be naturally is '.', due to that being the CWD.  I'm dumb, dunno why i did '.' earlier.  oops!

it's been fixed now tho.